% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_summary.R
\encoding{UTF-8}
\name{calculate_mode}
\alias{calculate_mode}
\title{Calculate Mode}
\usage{
calculate_mode(x, ...)

## Default S3 method:
calculate_mode(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{An R object. Currently there are methods for numeric/logical vectors
and date, date-time and time interval objects. Complex vectors are allowed for trim = 0, only.}

\item{na.rm}{a logical evaluating to TRUE or FALSE indicating whether NA
values should be stripped before the computation proceeds.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the most frequent value.
}
\description{
calculate_mode is a generic function to compute the most
abundant element in a list.
}
\references{
https://stackoverflow.com/a/8189441
}
\seealso{
[mean()] [median()] [sd()] [summary()]
}
\keyword{statistics}
\keyword{super_summary}
