% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ids.R
\name{find_symbol}
\alias{find_symbol}
\title{find_symbol}
\arguments{
\item{id}{a string for a gene id to match to gene symbol or aliases}

\item{annotation}{a list of gene symbols named by ids such as those provided by bioconductor AnnotationDbi packages}

\item{n}{number of symbols to list (maximum). Defaults to first. Set to Inf to display all.}
}
\description{
find gene symbol from annotation data
}
\examples{

#load data to identify genes from IDs
if(!require("org.At.tair.db")){
  install.packages("BiocManager")
  BiocManager::install("org.At.tair.db")
  library("org.At.tair.db")
}
annotation <- as.list(org.At.tairSYMBOL)
id <- "AT3G11260"
find_symbol(id, annotation)

#
ids <- c("AT3G11260", "AT2G17950")
lapply(ids, find_symbol, annotation)

# show multiple gene symbols
lapply(ids, find_symbol, annotation, n = 2)
# show all gene symbols
lapply(ids, find_symbol, annotation, n = Inf)

}
\seealso{
find_symbol
}
\keyword{annotation}
\keyword{bioconductor}
\keyword{bioinformatics}
\keyword{database}
\keyword{gene}
