% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_colour_hue.R
\name{gg_colour_hue}
\alias{gg_colour_hue}
\title{gg_colour_hue}
\arguments{
\item{n}{the number of colors (≥ 1) to be in the palette.}
}
\description{
emulate colour scale for ggplots
}
\examples{

# example plot
n <- 5
x_mean <- rnorm(n, 1000, 500)
y_mean <- rnorm(n, 1000, 500)
x_vals <- sapply(x_mean, function(x) rnorm(100, x, 50))
y_vals <- sapply(y_mean, function(y) rnorm(100, y, 50))
plot(x_vals, y_vals, col = colourscale(n, brewer.pal(11, "Set3"))[unlist(lapply(1:n, function(x) rep(x, 100)))], pch = 20,
     legend = legend("topleft", title = "colours", legend = 1:n, col = colourscale(n, brewer.pal(11, "Set3")), pch = 20),
     main = "title", xlab = "x-axis", ylab = "y-axis")

}
\keyword{annotation}
\keyword{graphics}
\keyword{legend}
\keyword{plot}
\keyword{title}
