% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot}
\alias{plot}
\title{plot}
\arguments{
\item{...}{parameters to pass to plot}

\item{legend}{plot legend}

\item{title}{plot title}

\item{mtext}{plot mtext to add}
}
\description{
plot with legend and title, mtext added
}
\examples{

# example plot
n <- 5
x_mean <- rnorm(n, 1000, 500)
y_mean <- rnorm(n, 1000, 500)
x_vals <- sapply(x_mean, function(x) rnorm(100, x, 50))
y_vals <- sapply(y_mean, function(y) rnorm(100, y, 50))
plot(x_vals, y_vals, col = colourscale(n, brewer.pal(11, "Set3"))[unlist(lapply(1:n, function(x) rep(x, 100)))], pch = 20,
     legend = legend("topleft", title = "colours", legend = 1:n, col = colourscale(n, brewer.pal(11, "Set3")), pch = 20),
     main = "title", xlab = "x-axis", ylab = "y-axis")

}
\keyword{annotation}
\keyword{graphics}
\keyword{legend}
\keyword{plot}
\keyword{title}
