% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{revcomp}
\alias{revcomp}
\alias{revcomp.factor}
\alias{revcomp.character}
\alias{revcomp.list}
\alias{revcomp.default}
\title{revcomp}
\usage{
\method{revcomp}{factor}(seq, upper.case = NULL, complement = TRUE,
  reverse = TRUE, RNA.out = FALSE)

\method{revcomp}{character}(seq, upper.case = NULL, complement = TRUE,
  reverse = TRUE, RNA.out = FALSE)

\method{revcomp}{list}(seq, upper.case = NULL, complement = TRUE,
  reverse = TRUE, RNA.out = FALSE)

\method{revcomp}{default}(seq, upper.case = NULL, complement = TRUE,
  reverse = TRUE, RNA.out = FALSE)
}
\arguments{
\item{seq}{string. A sequence of nucleotide bases (can handle ambiguous bases)}

\item{upper.case}{logical. Whether to force return of upper (TRUE) or lower (FALSE) case. Default to matching input case.}

\item{complement}{logical. Whether to compute the complementary sequence (defaults to TRUE).}

\item{reverse}{logical. Whether to reverse the order of the sequence (defaults to TRUE).}

\item{RNA.out}{logical. Whether to return sequences as DNA (TRUE) with T or as RNA (FALSE) with U (defaults to FAlSE).}
}
\value{
returns a string, vector, character, or list in the same class as the input sequences
}
\description{
compute the reverse-complement of a string of nucleotide bases
}
\examples{

# example plot
sequence <- "ATCG"
revcomp(sequence)
sequence <- "atcg"
revcomp(sequence)
sequence <- "AtgaNVVMKSvdbTC"
revcomp(sequence)
sequence <- c("ATACAG", "TATAAT", "TTGACA", "A", "c", "g", "nnn")
revcomp(sequence)
sequence <- list("ATACAG", "TATAAT", "TTGACA", "A", "c", "g", "nnn")
revcomp(sequence)
sequence <- factor("ATACAG", "TATAAT", "TTGACA", "A", "c", "g", "nnn")
revcomp(sequence)
sequence <- "ATCG"
revcomp(sequence, reverse = FALSE)
revcomp(sequence, complement = FALSE)
revcomp(sequence, upper.case = FALSE)

}
\keyword{bases}
\keyword{complement}
\keyword{molecular-biology}
\keyword{sequence}
\keyword{strings}
