% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitstrings}
\alias{splitstrings}
\title{plot}
\arguments{
\item{strings}{vector or list of strings to split}

\item{n}{indicies of each string to extract (and paste back together)}

\item{split}{character to split by}
}
\description{
plot with legend and title, mtext added
}
\examples{

# example string
string <- c("1-4-5", "4-6-3", "1-0-3")
splitstrings(string, 2)

}
\keyword{annotation}
\keyword{graphics}
\keyword{legend}
\keyword{plot}
\keyword{title}
